Compare matrix_to_operator, number and length of terms
matrix_to_operator_1 seems to always produce terms of length 2 while matrix_to_operator_2 produces terms which are growing with j (length 5 when j = 50) (will probably be log(size(H)). The number of terms also seems to be slightly lower for matrix_to_operator_1. The biggest advantage with matrix_to_operator_2 is, thus, that the number of qubits required grows logarithmically instead of linearly with the size of the matrix.