Debugging (see description)
The script crashes when trying to save. This seems to be because the result object contains objects that can't be pickled (the local function wrap_callbacks in vqe_override). The problem seems to be in result['correct'] when using gp_minimize. Here's the traceback:

AttributeError: Can't pickle local object 'VQE_override.vqe_run.<locals>.wrap_callbacks'